Добавил проверку привелегий пользователя перед получением доступа к эндпоинту

+++ src/api/v1/groups.py
@@ -5,6 +5,7 @@ from typing import Annotated
 from fastapi import APIRouter, Depends, HTTPException, Path, Body
 from fastapi.responses import JSONResponse
 from fastapi.encoders import jsonable_encoder
+from async_fastapi_jwt_auth import AuthJWT
 
 from schemas.entity import (
 	GroupDetailView,
@@ -13,6 +14,7 @@ from schemas.entity import (
 	GroupUpdate
 )
 from services.group import GroupService, get_group_service
+from services.authentication import AuthenticationService, get_authentication_service
 
 
 router = APIRouter()
@@ -27,8 +29,19 @@ router = APIRouter()
 )
 async def create_group(
 	group_create: Annotated[GroupCreate, Body(description='Шаблон для создания группы')],
-	group_service: GroupService = Depends(get_group_service)
+	group_service: GroupService = Depends(get_group_service),
+	authorize_service: AuthJWT = Depends(),
+	authentication_service: AuthenticationService = Depends(get_authentication_service)
 ) -> GroupDetailView:
+	await authorize_service.jwt_required()
+	is_authorized = await authentication_service.required_permissions(
+		await authorize_service.get_jwt_subject(),
+		['groups.create_group']
+	)
+
+	if not is_authorized:
+		raise HTTPException(status_code=HTTPStatus.FORBIDDEN, detail='Not enough rights')
+
 	group_create_encoded = jsonable_encoder(group_create)
 	group = await group_service.create_group(group_create_encoded)
 	if not group:
@@ -47,8 +60,19 @@ async def create_group(
 	response_description='Имена всех групп в базе данных'
 )
 async def read_groups(
-	group_service: GroupService = Depends(get_group_service)
+	group_service: GroupService = Depends(get_group_service),
+	authorize_service: AuthJWT = Depends(),
+	authentication_service: AuthenticationService = Depends(get_authentication_service)
 ) -> list[GroupShortView]:
+	await authorize_service.jwt_required()
+	is_authorized = await authentication_service.required_permissions(
+		await authorize_service.get_jwt_subject(),
+		['groups.read_groups']
+	)
+
+	if not is_authorized:
+		raise HTTPException(status_code=HTTPStatus.FORBIDDEN, detail='Not enough rights')
+
 	return await group_service.read_groups()
 
 
@@ -62,8 +86,19 @@ async def read_groups(
 async def update_group(
 	group_id: Annotated[UUID, Path(description='Идентификатор группы')],
 	group_update: Annotated[GroupUpdate, Body(description='Шаблон для изменения группы')],
-	group_service: GroupService = Depends(get_group_service)
+	group_service: GroupService = Depends(get_group_service),
+	authorize_service: AuthJWT = Depends(),
+	authentication_service: AuthenticationService = Depends(get_authentication_service)
 ) -> GroupDetailView:
+	await authorize_service.jwt_required()
+	is_authorized = await authentication_service.required_permissions(
+		await authorize_service.get_jwt_subject(),
+		['groups.update_group']
+	)
+
+	if not is_authorized:
+		raise HTTPException(status_code=HTTPStatus.FORBIDDEN, detail='Not enough rights')
+
 	group_update_encoded = jsonable_encoder(group_update)
 	group = await group_service.update_group(group_id, group_update_encoded)
 	if not group:
@@ -82,8 +117,19 @@ async def update_group(
 )
 async def delete_group(
 	group_id: Annotated[UUID, Path(description='Идентификатор группы')],
-	group_service: GroupService = Depends(get_group_service)
+	group_service: GroupService = Depends(get_group_service),
+	authorize_service: AuthJWT = Depends(),
+	authentication_service: AuthenticationService = Depends(get_authentication_service)
 ) -> JSONResponse:
+	await authorize_service.jwt_required()
+	is_authorized = await authentication_service.required_permissions(
+		await authorize_service.get_jwt_subject(),
+		['groups.delete_group']
+	)
+
+	if not is_authorized:
+		raise HTTPException(status_code=HTTPStatus.FORBIDDEN, detail='Not enough rights')
+
 	group_id = await group_service.delete_group(group_id)
 
 	if not group_id:
